+++ src/Paillave.Etl.EntityFrameworkCore.Tests/Entities/Blog.cs
@@ -1,0 +1,7 @@
+﻿namespace Paillave.Etl.EntityFrameworkCore.Tests.Entities;
+
+public class Blog
+{
+    public int BlogId { get; set; }
+    public string Name { get; set; }
+}
