+++ src/Paillave.Etl.EntityFrameworkCore.Tests/Entities/Post.cs
@@ -1,0 +1,7 @@
+﻿namespace Paillave.Etl.EntityFrameworkCore.Tests.Entities;
+
+public class Post
+{
+    public int PostId { get; set; }
+    public string Name { get; set; }
+}
